+++ nxtree/appinfo/routes.php
@@ -7,6 +7,7 @@ return [
         ['name' => 'page#index', 'url' => '/', 'verb' => 'GET'],
         ['name' => 'tree#index', 'url' => '/trees', 'verb' => 'GET'],
         ['name' => 'tree#create', 'url' => '/trees', 'verb' => 'POST'],
+        ['name' => 'tree#import', 'url' => '/import', 'verb' => 'POST'],
         ['name' => 'tree#show', 'url' => '/trees/{treeId}', 'verb' => 'GET'],
     ],
 ];